requirements.txt
flask==2.3.3
langchain==0.0.335
langchain-huggingface
langchain-community
pandas==2.1.0
python-dotenv==1.0.5
requests==2.31.0
huggingface-hub==0.19.0
Werkzeug==2.3.7
gunicorn==21.2.0
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4 